[FE] REFACTOR: 첫 로그인 시, 유저 정보 세팅 모달 리팩토링 #51
