+++ frontend/src/components/modules/UserInfoSettingModal/UserInfoSettingModal.tsx
@@ -9,6 +9,7 @@ import ModalTitle from "../ModalTitle";
 
 interface Props {
   onClose?: () => void;
+  onSubmit?: () => void;
 }
 
 const Wrapper = styled(Board).attrs((props) => {
@@ -22,22 +23,25 @@ const Wrapper = styled(Board).attrs((props) => {
   };
 })``;
 
-const UserInfoSettingModal = ({ onClose }: Props) => {
+const SubmitButton = styled(Button).attrs({
+  width: "50%",
+  height: "16%",
+  boxShadow: true,
+  fontSize: "2rem",
+})``;
+
+const UserInfoSettingModal = ({ onClose, onSubmit }: Props) => {
   return (
-    <>
-      <Modal width="30%" height="60%" animation>
-        <ModalTitle onClose={onClose} fontSize="3rem">
-          프로필설정
-        </ModalTitle>
-        <Wrapper>
-          <Avatar width="15vw" height="15vw" upload />
-          <Input height="15%" width="70%" borderRadius padding="1rem" />
-          <Button width="50%" height="16%" boxShadow fontSize="2rem">
-            설정완료
-          </Button>
-        </Wrapper>
-      </Modal>
-    </>
+    <Modal width="30%" height="60%" animation>
+      <ModalTitle onClose={onClose} fontSize="3rem">
+        프로필설정
+      </ModalTitle>
+      <Wrapper>
+        <Avatar width="15vw" height="15vw" upload />
+        <Input height="15%" width="70%" borderRadius padding="1rem" />
+        <SubmitButton onClick={onSubmit}>설정완료</SubmitButton>
+      </Wrapper>
+    </Modal>
   );
 };
